feat: add support to active record relations
You can now type `ActiveRecord->hasMany()` and `ActiveRecord->hasOne()` with
active record template params.

When you have the code

```php
/** @return ActiveQuery<Relation> */
public function getRelation(): ActiveQuery
{
	return $this->hasOne(Relation::class, ['id' => 'id']);
}
```

After resolving the query with `one()` or `all()` psalm will now know you have
an instance of the `Relation` class or `array<Relation>` if you call
`all()`. This prevents you from having to use `assert` or add dock blocks to
tell psalm this is a instance of a class not an `ActiveRecordInterface`