feat: add type resolution for properties via getters and setters
Psalm will now know the type of magic properties via the getters and getters on
a class that extends `BaseObject`.

```php
function getTest(): int {}
```

This snippet will new resolve the property `test` to be the type of `int`

```php
/** @return ActiveQuery<Test> */
function getTest(): ActiveQuery {}
```

Also active query classes are resolved to there template type. The above snippet
will resolve a property `test` to be the class type `Test`