feat: add multiple return type from ActiveRecord
You now need to add if the active record return type is multiple or not. You can
define the return type like.

```
/** @return ActiveRecord<User, true> */
```

This will define that all magic getters will return an `array<array-key, User>`
if false is passed then the magic getter is a `hasOne` relation and will return
a single user

Co-authored-by: StuTheWebGuy <stucarter18@gmail.com>