feat: add BaseYii stub for getting class-string return types
The plugin will now detect the return type of `Yii::createObject` from a class
string without needing to doc block the code

```php
$command = Yii::createObject(\yii\db\Command::class);
```

Psalm will now detect that command will be in instance of `yii\db\Command`. It
will also work with as array parameter with a `class` key.

```php
$command = Yii::createObject(['class' => \yii\db\Command::class]);
```

Co-authored-by: Anton <info@ensostudio.ru>